Modifying the GCN model, including modifying the weight initialization method, deepening the model, and transferring data to the GPU
